Move fqn to utilities module
